chat-backend/chat_backend/requirements.txt
asgiref==3.7.2
celery==5.4.0
channels==4.0.0
Django==4.2.6
djangorestframework==3.15.1
daphne==4.1.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
minio==7.2.0

